Add background image pattern to home header
This adds a little visual style to the header with a background pattern since there is not much going on with the solid color. It gives it a little visual flair.

The pattern in use is Japanese Sayagata by Olga Libby from Subtle
Patterns.
(https://www.toptal.com/designers/subtlepatterns/japanese-sayagata/)